claude/requirements.txt
Anthropic
requests
Python-dotenv